# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Helpdesk Ticket Management API",
        "description": "A simple API to manage helpdesk tickets."
    },
    "servers": [
        {
            "url": "http://localhost:3000/api"
        }
    ],
    "paths": {
        "/tickets": {
            "get": {
                "summary": "Get all tickets",
                "description": "Retrieve a list of all tickets with optional filtering and sorting.",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter tickets by status.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "accepted",
                                "resolved",
                                "rejected"
                            ]
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort tickets by the specified field (createdAt or updatedAt).",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tickets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new ticket",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TicketInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "put": {
                "summary": "Update a ticket's status or information",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID of the ticket to update.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TicketInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Ticket": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "title": {
                        "type": "string",
                        "example": "Login Issue"
                    },
                    "description": {
                        "type": "string",
                        "example": "Unable to log into the portal."
                    },
                    "contactInfo": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "accepted",
                            "resolved",
                            "rejected"
                        ],
                        "example": "pending"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-21T14:48:00.000Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-21T15:00:00.000Z"
                    }
                }
            },
            "TicketInput": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Login Issue"
                    },
                    "description": {
                        "type": "string",
                        "example": "Unable to log into the portal."
                    },
                    "contactInfo": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "accepted",
                            "resolved",
                            "rejected"
                        ],
                        "example": "pending"
                    }
                }
            }
        }
    }
}